backend/requirements.txt
asgi-redis==1.1.0
asgiref==3.6.0
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
constantly==15.1.0
cryptography==41.0.4
daphne==4.0.0
Django==4.2.4
django-browser-reload==1.11.0
django-extensions==3.2.3
django-mysql==4.11.0
django-sslserver==0.22
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Faker==19.10.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
MarkupSafe==2.1.3
msgpack==1.0.6
msgpack-python==0.5.6
Pillow==10.0.0
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
Werkzeug==2.3.7
whitenoise==6.5.0
zope.interface==6.0
